// libs/features/contributor/src/lib/shared/enums/document-transaction-id.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum DocumentTransactionId {
  REGISTER_CONTRIBUTOR = 'REGISTER_CONTRIBUTOR',
  CHANGE_ENGAGEMENT = 'CHANGE_ENGAGEMENT',
  MANAGE_WAGE = 'MANAGE_WAGE',
  TERMINATE_CONTRIBUTOR = 'TERMINATE_ENGAGEMENT',
  REGISTER_SECONDMENT = 'REGISTER_SECONDMENT',
  REGISTER_STUDY_LEAVE = 'REGISTER_STUDY_LEAVE',
  CANCEL_CONTRIBUTOR = 'CANCEL_ENGAGEMENT',
  ADD_SECONDED = 'REGISTER_SECONDED',
  TRANSFER_ENGAGEMENT = 'TRANSFER_ENGAGEMENT',
  TRANSFER_ALL_ENGAGEMENT = 'TRANSFER_ALL_ENGAGEMENT',
  REGISTER_VIC_CONTRIBUTOR = 'REGISTER_VIC_CONTRIBUTOR',
  MANAGE_VIC_WAGE = 'MANAGE_VIC_WAGE',
  TERMINATE_VIC = 'TERMINATE_VIC_ENGAGEMENT',
  CANCEL_VIC = 'CANCEL_VIC_ENGAGEMENT',
  ADD_CONTRACT = 'REGISTER_CONTRACT',
  CANCEL_CONTRACT = 'CANCEL_CONTRACT',
  CANCEL_ENGAGEMENT_VIOLATION = 'REGISTER_CANCEL_ENGAGEMENT_VIOLATION',
  CHANGE_ENGAGEMENT_VIOLATION = 'REGISTER_CHANGE_ENGAGEMENT_VIOLATION',
  TERMINATE_ENGAGEMENT_VIOLATION = 'REGISTER_TERMINATE_ENGAGEMENT_VIOLATION',
  CONTRACT_DOCUMENT = 'CONTRACT_AUTHENTICATION_DOCUMENT',
  ADD_AUTHORIZATION = 'ADD_AUTHORIZATION_EXT',
  MAINTAIN_COVERAGE = 'MAINTAIN_COVERAGE',
  MODIFY_PERSONAL_DETAILS = 'MODIFY_PERSONAL_DETAILS',
  ADD_FAMILY_DETAILS = 'ADD_FAMILY_DETAILS',
  MODIFY_FAMILY_DETAILS = 'MODIFY_FAMILY_DETAILS',
  UPDATE_BORDER_NUMBER = 'UPDATE_BORDER_NUMBER',
  UPDATE_IQAMA_NUMBER = 'UPDATE_IQAMA_NUMBER',
  UPLOAD_DOCUMENTS = 'UPLOAD_DOCUMENTS',
  ADD_ENGAGEMENT_E_INSPECTION = 'ADD_ENGAGEMENT_E_INSPECTION',
  REACTIVATE_ENGAGEMENT = 'REACTIVATE_ENGAGEMENT',
  REACTIVATE_VIC_ENGAGEMENT = 'REACTIVATE_VIC_ENGAGEMENT',
  ADD_NIN = 'ADD_NIN',
  EDIT_NIN = 'EDIT_NIN',
  ENTER_RPA_AGGREGATION = 'ENTER_RPA_FIRST_SCHEME',
  ENTER_RPA_AGGREGATION_PPA = 'ENTER_RPA_LAST_SCHEME'
}